feat: vis-encode all path characters except visible ASCII
This allows archive input files to include Unicode characters.
It should also support the more esoteric corners of the ASCII set
(e.g. backslashes, newlines, control codes), if the user can find a way
to convince Bazel to use such a file. This latter feature probably isn't
a good idea to actually use; but it is nice to have confidence that this
ruleset won't choke if it comes up.

Generating an archive with non-ASCII characters reports a diagnostic,
however this does not appear to correspond to a real issue — the file
is still placed into the archive with the expected path and contents.

```
INFO: From Tar lib/tests/tar/7.tar:
tar: lib/tests/tar/srcdir/Unicode® support?🤞: Can't translate pathname 'lib/tests/tar/srcdir/Unicode® support?🤞' to UTF-8
```
Encoding of 7-bit ASCII could be performed in Bazel, however Starlark does not provide
access to a string's bytes, only its codepoints, at least until the spec and implementation work
to get a [`bytes` type](https://github.com/bazelbuild/starlark/issues/112) lands.
So a second, out-of-process build-time pass is needed. Since a separate pass
is needed anyways, we move most of the encoding responsibility to this external script
(except cases that it cannot handle without interfering with the mtree file format: newline, space, and backslash).

We use `gawk` for this second pass. A [prior attempt](https://github.com/bazel-contrib/bazel-lib/pull/795)
at this feature had done a similar external pass with the `vis` utility, but has
not been accepted because the `vis` tool is not expected to be available everywhere.
`gawk` is provided through a [BCR module of the same
name](https://registry.bazel.build/modules/gawk).

The extended feature set of GNU Awk is necessary to implement these
scripts succinctly; POSIX `awk` that might be assumed to be present on
most systems is insufficient. At the very least,
it is difficult to do byte-by-byte encoding without
`--characters-as-bytes` nor efficient, one-pass, table-based replacement without
4-arg `split`.

We also take this opportunity to canonicalize the paths derived
from user-provided mtree specs when computing the unused
inputs set. This should ensure that the comparison is exact,
which substantially reduces the risks associated with this feature
and should enable it to become a default.

Fixes https://github.com/bazel-contrib/bazel-lib/issues/794