Satisfy terminal-whitespace pre-commit
Rather than representing the gap corresponding to the "Delete" character
with placeholder whitespace, we just end the line then and there. This
is fine in this case because the gap comes at the end of the sequence,
and logically the "Delete" character isn't really in the character set
of interest, it just happens to be immediately adjacent and fills out
the block to a nicely padded number.